+++ scripts/superdesk-search-providers/constants.js
@@ -1,0 +1,10 @@
+export const providerTypes = {
+    aapmm: 'AAP Multimedia',
+    paimg: 'PA Images',
+    // quick fix to have several scanpix instance (needed for SDNTB-237)
+    // FIXME: temporary fix, need to be refactored (SD-4448)
+    'scanpix(ntbtema)': 'ScanPix (ntbtema)',
+    'scanpix(ntbkultur)': 'ScanPix (ntbkultur)',
+    'scanpix(desk)': 'ScanPix (desk)',
+    'scanpix(npk)': 'ScanPix (npk)',
+};
